+++ EnhanceQoL/General/functions.lua
@@ -11,6 +11,13 @@ local GetBagItem = C_TooltipInfo.GetBagItem
 local IsEquippableItemFn = C_Item.IsEquippableItem
 addon.functions = addon.functions or {}
 
+function addon.functions.IsMinigameClientScene(sceneType)
+	if Enum and Enum.ClientSceneType and Enum.ClientSceneType.MinigameSceneType ~= nil then
+		return sceneType == Enum.ClientSceneType.MinigameSceneType
+	end
+	return sceneType == 1
+end
+
 local UnitHealth, UnitHealthMax = UnitHealth, UnitHealthMax
 local UnitPower, UnitPowerMax = UnitPower, UnitPowerMax
 local UnitHealthPercent = UnitHealthPercent